Separated internal MPMA functions to another file, enhanced style and added new safeguards
