+++ counterpartylib/lib/messages/send.py
@@ -27,7 +27,7 @@ def initialise (db):
     #   SQLite can’t do `ALTER TABLE IF COLUMN NOT EXISTS` nor can drop UNIQUE constraints
     if 'msg_index' not in columns:
         if 'memo' not in columns:
-            cursor.execute('''CREATE TABLE new_sends(
+            cursor.execute('''CREATE TABLE IF NOT EXISTS new_sends(
                               tx_index INTEGER,
                               tx_hash TEXT,
                               block_index INTEGER,
@@ -45,7 +45,7 @@ def initialise (db):
                 SELECT tx_index, tx_hash, block_index, source, destination, asset, quantity, status
                 FROM sends''', {})
         else:
-            cursor.execute('''CREATE TABLE new_sends(
+            cursor.execute('''CREATE TABLE IF NOT EXISTS new_sends(
                   tx_index INTEGER,
                   tx_hash TEXT,
                   block_index INTEGER,
@@ -96,9 +96,6 @@ def unpack(db, message, block_index):
 def validate (db, source, destination, asset, quantity, block_index):
     return send1.validate(db, source, destination, asset, quantity, block_index)
 
-def flat(z):
-    return [x for x in z]
-
 def compose (db, source, destination, asset, quantity, memo=None, memo_is_hex=False, use_enhanced_send=None):
     # special case - enhanced_send replaces send by default when it is enabled
     #   but it can be explicitly disabled with an API parameter
@@ -106,12 +103,38 @@ def compose (db, source, destination, asset, quantity, memo=None, memo_is_hex=Fa
         # Another special case, if destination, asset and quantity are arrays, it's an MPMA send
         if isinstance(destination, list) and isinstance(asset, list) and isinstance(quantity, list):
             if util.enabled('mpma_sends'):
-                if isinstance(memo, list) and isinstance(memo_is_hex, list):
-                    return mpma.compose(db, source, flat(zip(asset, destination, quantity, memo, memo_is_hex)), None, None)
-                elif isinstance(memo, dict) and isinstance(memo_is_hex, dict):
-                    return mpma.compose(db, source, flat(zip(asset, destination, quantity, memo['list'], memo_is_hex['list'])), memo['msg_wide'], memo_is_hex['msg_wide'])
+                if len(destination) == len(asset) and len(asset) == len(quantity):
+                    # Sending memos in a MPMA message can be done by several approaches:
+                    # 1. Send a list of memos, there must be one for each send and they correspond to the sends by index
+                    #   - In this case memo_is_hex should be a list with the same cardinality
+                    # 2. Send a dict with the message specific memos and the message wide memo (same for the hex specifier):
+                    #   - Each dict should have 2 members:
+                    #     + list: same as case (1). An array that specifies the memo for each send
+                    #     + msg_wide: the memo for the whole message. This memo will be used for sends that don't have a memo specified. Same as in (3)
+                    # 3. Send one memo (either bytes or string) and True/False in memo_is_hex. This will be interpreted as a message wide memo.
+                    if isinstance(memo, list) and isinstance(memo_is_hex, list):
+                        # (1) implemented here
+                        if len(memo) != len(memo_is_hex):
+                            raise exceptions.ComposeError('memo and memo_is_hex lists should have the same length')
+                        elif len(memo) != len(destination):
+                            raise exceptions.ComposeError('memo/memo_is_hex lists should have the same length as sends')
+
+                        return mpma.compose(db, source, util.flat(zip(asset, destination, quantity, memo, memo_is_hex)), None, None)
+                    elif isinstance(memo, dict) and isinstance(memo_is_hex, dict):
+                        # (2) implemented here
+                        if not('list' in memo and 'list' in memo_is_hex and 'msg_wide' in memo and 'msg_wide' in memo_is_hex):
+                            raise exceptions.ComposeError('when specifying memo/memo_is_hex as a dict, they must contain keys "list" and "msg_wide"')
+                        elif len(memo['list']) != len(memo_is_hex['list']):
+                            raise exceptions.ComposeError('length of memo.list and memo_is_hex.list must be equal')
+                        elif len(memo['list']) != len(destination):
+                            raise exceptions.ComposeError('length of memo.list/memo_is_hex.list must be equal to the amount of sends')
+
+                        return mpma.compose(db, source, util.flat(zip(asset, destination, quantity, memo['list'], memo_is_hex['list'])), memo['msg_wide'], memo_is_hex['msg_wide'])
+                    else:
+                        # (3) the default case
+                        return mpma.compose(db, source, util.flat(zip(asset, destination, quantity)), memo, memo_is_hex)
                 else:
-                    return mpma.compose(db, source, flat(zip(asset, destination, quantity)), memo, memo_is_hex)
+                    raise exceptions.ComposeError('destination, asset and quantity arrays must have the same amount of elements')
             else:
                 raise exceptions.ComposeError('mpma sends are not enabled')
         elif use_enhanced_send is None or use_enhanced_send == True:
